refactor core agent types
