+++ packages/core/src/agent/agent.ts
@@ -22,103 +22,20 @@ import type {
   ToolCallContext,
 } from "../tool/tool";
 import { ToolSet } from "../tool/tool-set";
-import type { VectorFilter, VectorSearchIndex, VectorSearchResult } from "../vector-store";
+import type { VectorSearchIndex } from "../vector-store";
 import type { PromptHook } from "./hooks";
 import { normalizeAgentId } from "./ids";
 import { PromptRequest } from "./request";
+import type {
+  AgentEventStoreRegistration,
+  AgentOptions,
+  AgentToolOptions,
+  DynamicContextRegistration,
+  DynamicToolRegistration,
+} from "./types";
 import { isStreamingCompletionModel } from "./utils";
 
-export type AgentOptions<M extends CompletionModel = CompletionModel> = {
-  id: string;
-  name?: string | undefined;
-  description?: string | undefined;
-  model: M;
-  instructions?: string | undefined;
-  staticContext?: Document[];
-  temperature?: number | undefined;
-  maxTokens?: number | undefined;
-  additionalParams?: JsonValue | undefined;
-  toolSet?: ToolSet | undefined;
-  toolChoice?: ToolChoice | undefined;
-  defaultMaxTurns?: number | undefined;
-  hook?: PromptHook | undefined;
-  outputSchema?: JsonObject | undefined;
-  observers?: AgentObserverRegistration[] | undefined;
-  approvals?: ToolApprovalsOptions | undefined;
-  dynamicContexts?: DynamicContextRegistration[] | undefined;
-  dynamicTools?: DynamicToolRegistration[] | undefined;
-  middlewares?: AgentMiddleware[] | undefined;
-  /**
-   * @deprecated Use `middlewares` instead.
-   */
-  toolMiddlewares?: AgentMiddleware[] | undefined;
-  memory?: MemoryRegistration | undefined;
-  eventStore?: AgentEventStoreRegistration | undefined;
-};
-
 export const DEFAULT_MAX_TURNS = 20;
-
-export type AgentToolOptions = {
-  name: string;
-  description?: string | undefined;
-  maxTurns?: number | undefined;
-  stream?: boolean | undefined;
-};
-
-export type AgentEventStoreInclude = "all" | "agent_tool_events";
-
-export type AgentEventStoreOptions = {
-  include?: AgentEventStoreInclude | undefined;
-};
-
-export type AgentEventAppendInput = {
-  runId: string;
-  agentId: string;
-  agentName?: string | undefined;
-  turn?: number | undefined;
-  toolName?: string | undefined;
-  toolCallId?: string | undefined;
-  internalCallId?: string | undefined;
-  event: unknown;
-};
-
-export type AgentEventRecord = AgentEventAppendInput & {
-  createdAt?: Date | undefined;
-};
-
-export interface AgentEventStore {
-  append(input: AgentEventAppendInput): Promise<void>;
-  load(runId: string): Promise<AgentEventRecord[]>;
-  clear?(runId: string): Promise<void>;
-}
-
-export type AgentEventStoreRegistration = {
-  store: AgentEventStore;
-  options: Required<AgentEventStoreOptions>;
-};
-
-export type DynamicContextOptions<T = unknown> = {
-  topK: number;
-  threshold?: number | undefined;
-  filter?: VectorFilter | undefined;
-  format?: ((result: VectorSearchResult<T>) => Document) | undefined;
-};
-
-export type DynamicContextRegistration<T = unknown> = {
-  index: VectorSearchIndex<T>;
-  options: DynamicContextOptions<T>;
-};
-
-export type DynamicToolOptions = {
-  topK: number;
-  threshold?: number | undefined;
-  filter?: VectorFilter | undefined;
-};
-
-export type DynamicToolRegistration = {
-  index: VectorSearchIndex<ToolSearchDocument>;
-  options: DynamicToolOptions;
-};
 
 export class Agent<M extends CompletionModel = CompletionModel> {
   readonly id: string;
