+++ packages/core/src/internal/agent.ts
@@ -1,1 +1,2 @@
 export * from "../agent/agent";
+export type * from "../agent/types";
